+++ mtdata/__init__.py
@@ -4,7 +4,7 @@
 # Created: 4/4/20
 
 
-__version__ = '0.4.2'
+__version__ = '0.4.3'
 __description__ = 'mtdata is a tool to download datasets for machine translation'
 __author__ = 'Thamme Gowda'
 
@@ -26,6 +26,18 @@ pbar_man = enlighten.get_manager()
 
 class MTDataException(Exception):
     pass
+
+
+class MTDataUserError(MTDataException):
+    """
+    This exception is for the cases where printing the whole stack trace is bad UI
+    and we want to show a user-friendly message. https://github.com/thammegowda/mtdata/issues/162
+    """
+    def __init__(self, msg, exitcode=1, *args):
+        super().__init__(*args)
+        self.msg = msg
+        self.exitcode = exitcode
+
 
 class Defaults:
     FILE_LOCK_TIMEOUT = 2 * 60 * 60  # 2 hours