+++ mtdata/utils.py
@@ -90,7 +90,7 @@ class IO:
     def get_lines(cls, path, col=0, delim='\t', line_mapper=None, newline_fix=True):
         with cls.reader(path) as inp:
             if newline_fix and delim != '\r':
-                inp = (line.replace(b'\r', b'') for line in inp)
+                inp = (line.replace('\r', '') for line in inp)
             if col >= 0:
                 inp = (line.split(delim)[col].strip() for line in inp)
             if line_mapper:
@@ -181,7 +181,7 @@ class TarPath(ArchivedPath):
         self.ext_dir = self.extract()
         matches = list(self.ext_dir.glob(self.name))
         if len(matches) != 1:
-            raise Exception(f'expected to find exactly one path inside tarball, but found {matches}')
+            raise Exception(f'expected to find exactly one path inside tarball @ {self.ext_dir}/{self.name}, but found {matches}')
         self.child = matches[0]
         self.open = self.child.open